[FEAT] #76 - 일정 삭제 버튼 이동 (AddEditFragment -> PlanDetailsFragment)
