+++ app/src/main/java/com/ivyclub/contact/ui/main/plan_details/PlanDetailsFragment.kt
@@ -2,6 +2,7 @@ package com.ivyclub.contact.ui.main.plan_details
 
 import android.os.Bundle
 import android.view.View
+import android.widget.Toast
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
@@ -9,6 +10,7 @@ import com.ivyclub.contact.R
 import com.ivyclub.contact.databinding.FragmentPlanDetailsBinding
 import com.ivyclub.contact.util.BaseFragment
 import com.ivyclub.contact.util.setFriendChips
+import com.ivyclub.contact.util.showAlertDialog
 import dagger.hilt.android.AndroidEntryPoint
 import java.text.SimpleDateFormat
 
@@ -31,18 +33,38 @@ class PlanDetailsFragment :
     }
 
     private fun setEditPlanButton() {
-        binding.ivBtnEditPlan.setOnClickListener {
-            findNavController().navigate(
-                PlanDetailsFragmentDirections.actionPlanDetailsFragmentToAddEditFragment(
-                    args.planId
+        with(binding) {
+            binding.ivBtnEditPlan.setOnClickListener {
+                findNavController().navigate(
+                    PlanDetailsFragmentDirections.actionPlanDetailsFragmentToAddEditFragment(
+                        args.planId
+                    )
                 )
-            )
+            }
+            binding.ivBtnDeletePlan.setOnClickListener {
+                showDeletePlanDialog()
+            }
         }
+    }
+
+    private fun showDeletePlanDialog() {
+        context?.showAlertDialog(getString(R.string.ask_delete_plan), {
+            viewModel.deletePlan()
+        })
     }
 
     private fun setObservers() {
         viewModel.planParticipants.observe(viewLifecycleOwner) {
             binding.cgPlanParticipants.setFriendChips(it)
+        }
+
+        viewModel.toastMessage.observe(viewLifecycleOwner) {
+            if (context == null) return@observe
+            Toast.makeText(context, getString(it), Toast.LENGTH_SHORT).show()
+        }
+
+        viewModel.finishEvent.observe(viewLifecycleOwner) {
+            findNavController().popBackStack()
         }
     }
