+++ app/src/main/res/layout/fragment_add_edit_plan.xml
@@ -32,16 +32,6 @@
                 app:layout_constraintTop_toTopOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"/>
 
-            <ImageView
-                android:id="@+id/iv_btn_delete_plan"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:padding="4dp"
-                android:layout_margin="16dp"
-                android:src="@drawable/ic_delete"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintStart_toStartOf="parent"/>
-
             <EditText
                 android:id="@+id/et_plan_title"
                 android:layout_width="wrap_content"
